Fixes UI (integration) tests
Upgrades support library version
Migrate legacy tests to use ActivityTestRule
